+++ src/lib/types_test.rs
@@ -372,6 +372,7 @@ fn task_new() {
     assert!(task.command.is_none());
     assert!(task.disabled.is_none());
     assert!(task.private.is_none());
+    assert!(task.extend.is_none());
     assert!(task.watch.is_none());
     assert!(task.condition.is_none());
     assert!(task.condition_script.is_none());
@@ -470,6 +471,7 @@ fn task_extend_both_have_misc_data() {
         workspace: None,
         disabled: Some(true),
         private: Some(true),
+        extend: None,
         watch: Some(TaskWatchOptions::Boolean(true)),
         condition: None,
         condition_script: None,
@@ -505,6 +507,7 @@ fn task_extend_both_have_misc_data() {
     assert!(base.workspace.is_none());
     assert!(base.disabled.is_some());
     assert!(base.private.is_some());
+    assert!(base.extend.is_none());
     assert!(base.watch.is_some());
     assert!(base.condition.is_none());
     assert!(base.condition_script.is_none());
@@ -555,6 +558,7 @@ fn task_extend_extended_have_all_fields() {
         workspace: None,
         disabled: Some(false),
         private: Some(true),
+        extend: Some("base".to_string()),
         watch: Some(TaskWatchOptions::Boolean(true)),
         condition: None,
         condition_script: None,
@@ -592,6 +596,7 @@ fn task_extend_extended_have_all_fields() {
         workspace: Some(true),
         disabled: Some(true),
         private: Some(false),
+        extend: Some("extended".to_string()),
         watch: Some(TaskWatchOptions::Boolean(false)),
         condition: Some(TaskCondition {
             profiles: Some(vec!["development".to_string()]),
@@ -626,6 +631,7 @@ fn task_extend_extended_have_all_fields() {
             command: Some("test2".to_string()),
             disabled: Some(true),
             private: Some(false),
+            extend: Some("extended".to_string()),
             watch: Some(TaskWatchOptions::Boolean(false)),
             condition: Some(TaskCondition {
                 profiles: Some(vec!["development".to_string()]),
@@ -657,6 +663,7 @@ fn task_extend_extended_have_all_fields() {
             command: Some("test2".to_string()),
             disabled: Some(true),
             private: Some(false),
+            extend: Some("extended".to_string()),
             watch: Some(TaskWatchOptions::Boolean(false)),
             condition: Some(TaskCondition {
                 profiles: Some(vec!["development".to_string()]),
@@ -688,6 +695,7 @@ fn task_extend_extended_have_all_fields() {
             command: Some("test2".to_string()),
             disabled: Some(true),
             private: Some(false),
+            extend: Some("extended".to_string()),
             watch: Some(TaskWatchOptions::Boolean(false)),
             condition: Some(TaskCondition {
                 profiles: Some(vec!["development".to_string()]),
@@ -725,6 +733,7 @@ fn task_extend_extended_have_all_fields() {
     assert!(base.workspace.is_some());
     assert!(base.disabled.is_some());
     assert!(base.private.is_some());
+    assert!(base.extend.is_some());
     assert!(base.watch.is_some());
     assert!(base.condition.is_some());
     assert!(base.condition_script.is_some());
@@ -759,6 +768,7 @@ fn task_extend_extended_have_all_fields() {
     assert!(base.workspace.unwrap());
     assert!(base.disabled.unwrap());
     assert!(!base.private.unwrap());
+    assert_eq!(base.extend.unwrap(), "extended");
     assert_eq!(base.watch.unwrap(), TaskWatchOptions::Boolean(false));
     assert_eq!(base.condition_script.unwrap().len(), 1);
     assert!(!base.ignore_errors.unwrap());
@@ -803,6 +813,7 @@ fn task_extend_clear_with_no_data() {
         workspace: Some(false),
         disabled: Some(true),
         private: Some(false),
+        extend: Some("base".to_string()),
         watch: Some(TaskWatchOptions::Boolean(false)),
         condition: Some(TaskCondition {
             profiles: Some(vec!["development".to_string()]),
@@ -837,6 +848,7 @@ fn task_extend_clear_with_no_data() {
             command: Some("test2".to_string()),
             disabled: Some(true),
             private: Some(false),
+            extend: Some("base".to_string()),
             watch: Some(TaskWatchOptions::Boolean(false)),
             condition: Some(TaskCondition {
                 profiles: Some(vec!["development".to_string()]),
@@ -868,6 +880,7 @@ fn task_extend_clear_with_no_data() {
             command: Some("test2".to_string()),
             disabled: Some(true),
             private: Some(false),
+            extend: Some("base".to_string()),
             watch: Some(TaskWatchOptions::Boolean(false)),
             condition: Some(TaskCondition {
                 profiles: Some(vec!["development".to_string()]),
@@ -899,6 +912,7 @@ fn task_extend_clear_with_no_data() {
             command: Some("test2".to_string()),
             disabled: Some(true),
             private: Some(false),
+            extend: Some("base".to_string()),
             watch: Some(TaskWatchOptions::Boolean(false)),
             condition: Some(TaskCondition {
                 profiles: Some(vec!["development".to_string()]),
@@ -938,6 +952,7 @@ fn task_extend_clear_with_no_data() {
     assert!(base.workspace.is_none());
     assert!(base.disabled.is_none());
     assert!(base.private.is_none());
+    assert!(base.extend.is_none());
     assert!(base.watch.is_none());
     assert!(base.condition.is_none());
     assert!(base.condition_script.is_none());
@@ -978,6 +993,7 @@ fn task_extend_clear_with_all_data() {
         workspace: Some(true),
         disabled: Some(true),
         private: Some(false),
+        extend: Some("base".to_string()),
         watch: Some(TaskWatchOptions::Boolean(false)),
         condition: Some(TaskCondition {
             profiles: Some(vec!["development".to_string()]),
@@ -1012,6 +1028,7 @@ fn task_extend_clear_with_all_data() {
             command: Some("test2".to_string()),
             disabled: Some(true),
             private: Some(false),
+            extend: Some("base".to_string()),
             watch: Some(TaskWatchOptions::Boolean(false)),
             condition: Some(TaskCondition {
                 profiles: Some(vec!["development".to_string()]),
@@ -1043,6 +1060,7 @@ fn task_extend_clear_with_all_data() {
             command: Some("test2".to_string()),
             disabled: Some(true),
             private: Some(false),
+            extend: Some("base".to_string()),
             watch: Some(TaskWatchOptions::Boolean(false)),
             condition: Some(TaskCondition {
                 profiles: Some(vec!["development".to_string()]),
@@ -1074,6 +1092,7 @@ fn task_extend_clear_with_all_data() {
             command: Some("test2".to_string()),
             disabled: Some(true),
             private: Some(false),
+            extend: Some("base".to_string()),
             watch: Some(TaskWatchOptions::Boolean(false)),
             condition: Some(TaskCondition {
                 profiles: Some(vec!["development".to_string()]),
@@ -1110,6 +1129,7 @@ fn task_extend_clear_with_all_data() {
     assert!(base.workspace.is_some());
     assert!(base.disabled.is_some());
     assert!(base.private.is_some());
+    assert!(base.extend.is_some());
     assert!(base.watch.is_some());
     assert!(base.condition.is_some());
     assert!(base.condition_script.is_some());
@@ -1183,6 +1203,7 @@ fn task_get_normalized_task_undefined() {
         command: Some("command".to_string()),
         disabled: Some(false),
         private: Some(true),
+        extend: Some("base".to_string()),
         watch: Some(TaskWatchOptions::Boolean(true)),
         condition: None,
         condition_script: None,
@@ -1214,6 +1235,7 @@ fn task_get_normalized_task_undefined() {
     assert!(normalized_task.command.is_some());
     assert!(normalized_task.disabled.is_some());
     assert!(normalized_task.private.is_some());
+    assert!(normalized_task.extend.is_some());
     assert!(normalized_task.watch.is_some());
     assert!(normalized_task.condition.is_none());
     assert!(normalized_task.condition_script.is_none());
@@ -1250,6 +1272,7 @@ fn task_get_normalized_task_undefined() {
     assert!(!normalized_task.workspace.unwrap());
     assert!(!normalized_task.disabled.unwrap());
     assert!(normalized_task.private.unwrap());
+    assert_eq!(normalized_task.extend.unwrap(), "base");
     assert_eq!(
         normalized_task.watch.unwrap(),
         TaskWatchOptions::Boolean(true)
@@ -1294,6 +1317,7 @@ fn task_get_normalized_task_with_override_no_clear() {
         workspace: Some(true),
         disabled: Some(false),
         private: Some(true),
+        extend: Some("base".to_string()),
         watch: Some(TaskWatchOptions::Boolean(true)),
         condition: Some(TaskCondition {
             profiles: Some(vec!["development".to_string()]),
@@ -1324,6 +1348,7 @@ fn task_get_normalized_task_with_override_no_clear() {
             command: Some("linux_command".to_string()),
             disabled: Some(true),
             private: Some(false),
+            extend: Some("linux".to_string()),
             watch: Some(TaskWatchOptions::Boolean(false)),
             condition: Some(TaskCondition {
                 profiles: Some(vec!["development".to_string()]),
@@ -1368,6 +1393,7 @@ fn task_get_normalized_task_with_override_no_clear() {
     assert!(normalized_task.workspace.is_some());
     assert!(normalized_task.disabled.is_some());
     assert!(normalized_task.private.is_some());
+    assert!(normalized_task.extend.is_some());
     assert!(normalized_task.watch.is_some());
     assert!(normalized_task.condition.is_some());
     assert!(normalized_task.condition_script.is_some());
@@ -1402,6 +1428,7 @@ fn task_get_normalized_task_with_override_no_clear() {
     assert!(normalized_task.workspace.unwrap());
     assert!(normalized_task.disabled.unwrap());
     assert!(!normalized_task.private.unwrap());
+    assert_eq!(normalized_task.extend.unwrap(), "linux");
     assert_eq!(
         normalized_task.watch.unwrap(),
         TaskWatchOptions::Boolean(false)
@@ -1449,6 +1476,7 @@ fn task_get_normalized_task_with_override_clear_false() {
         workspace: Some(true),
         disabled: Some(false),
         private: Some(true),
+        extend: Some("base".to_string()),
         watch: Some(TaskWatchOptions::Boolean(true)),
         condition: Some(TaskCondition {
             profiles: Some(vec!["development".to_string()]),
@@ -1478,6 +1506,7 @@ fn task_get_normalized_task_with_override_clear_false() {
             command: Some("linux_command".to_string()),
             disabled: Some(true),
             private: Some(false),
+            extend: Some("linux".to_string()),
             watch: Some(TaskWatchOptions::Boolean(false)),
             condition: Some(TaskCondition {
                 profiles: Some(vec!["development".to_string()]),
@@ -1526,6 +1555,7 @@ fn task_get_normalized_task_with_override_clear_false() {
     assert!(normalized_task.workspace.is_some());
     assert!(normalized_task.disabled.is_some());
     assert!(normalized_task.private.is_some());
+    assert!(normalized_task.extend.is_some());
     assert!(normalized_task.watch.is_some());
     assert!(normalized_task.condition.is_some());
     assert!(normalized_task.condition_script.is_some());
@@ -1560,6 +1590,7 @@ fn task_get_normalized_task_with_override_clear_false() {
     assert!(normalized_task.workspace.unwrap());
     assert!(normalized_task.disabled.unwrap());
     assert!(!normalized_task.private.unwrap());
+    assert_eq!(normalized_task.extend.unwrap(), "linux");
     assert_eq!(
         normalized_task.watch.unwrap(),
         TaskWatchOptions::Boolean(false)
@@ -1602,6 +1633,7 @@ fn task_get_normalized_task_with_override_clear_false_partial_override() {
         command: Some("command".to_string()),
         disabled: Some(false),
         private: Some(true),
+        extend: Some("base".to_string()),
         watch: Some(TaskWatchOptions::Boolean(true)),
         condition: Some(TaskCondition {
             profiles: Some(vec!["development".to_string()]),
@@ -1635,6 +1667,7 @@ fn task_get_normalized_task_with_override_clear_false_partial_override() {
             command: None,
             disabled: None,
             private: None,
+            extend: None,
             watch: None,
             condition: None,
             condition_script: None,
@@ -1663,6 +1696,7 @@ fn task_get_normalized_task_with_override_clear_false_partial_override() {
     assert!(normalized_task.command.is_some());
     assert!(normalized_task.disabled.is_some());
     assert!(normalized_task.private.is_some());
+    assert!(normalized_task.extend.is_some());
     assert!(normalized_task.watch.is_some());
     assert!(normalized_task.condition.is_some());
     assert!(normalized_task.condition_script.is_some());
@@ -1696,6 +1730,7 @@ fn task_get_normalized_task_with_override_clear_false_partial_override() {
     assert_eq!(normalized_task.command.unwrap(), "command");
     assert!(!normalized_task.disabled.unwrap());
     assert!(normalized_task.private.unwrap());
+    assert_eq!(normalized_task.extend.unwrap(), "base");
     assert_eq!(
         normalized_task.watch.unwrap(),
         TaskWatchOptions::Boolean(true)
@@ -1733,6 +1768,7 @@ fn task_get_normalized_task_with_override_clear_true() {
         command: Some("command".to_string()),
         disabled: Some(false),
         private: Some(true),
+        extend: Some("base".to_string()),
         watch: Some(TaskWatchOptions::Boolean(true)),
         condition: Some(TaskCondition {
             profiles: Some(vec!["development".to_string()]),
@@ -1766,6 +1802,7 @@ fn task_get_normalized_task_with_override_clear_true() {
             command: None,
             disabled: None,
             private: None,
+            extend: None,
             watch: None,
             condition: None,
             condition_script: None,
@@ -1794,6 +1831,7 @@ fn task_get_normalized_task_with_override_clear_true() {
     assert!(normalized_task.command.is_none());
     assert!(normalized_task.disabled.is_none());
     assert!(normalized_task.private.is_none());
+    assert!(normalized_task.extend.is_none());
     assert!(normalized_task.watch.is_none());
     assert!(normalized_task.condition.is_none());
     assert!(normalized_task.condition_script.is_none());